Major changes - updated data binding option
New toption 'watchDataItems' is a new flag that tells the grid to watch
individual objects in the array for changes and updates the grid.
Also, in order to update the grid by reassigning the data array we need
to get the string name of the object on the scope which contains your
data.