Fetch templates with $http, run init() after
getTemplate() now return a promise. It is immediately resolved if the
template is local. If it is remote, the promise is resolved after the
template is fetched and put into the cache.  Also the $http request is
cached in $templateCache as well, so the same template should only be
fetched once no matter now many grids use it.

New function self.initTemplates(), calls getTemplate() for each of the template
types. Returns a promise via $q.all() that bundles all the getTemplate()
promises.

self.init() now returns a promise and defers execution until after
initTemplates() completes.