Fixes #399, making cell editing work again
This is an issue with how src/ng-input.js is handling the value of the ng-input directive. The current default template contains an $eval, which means that when ng-input.js runs a $parse on it, it gets the evaluated value of the cell itself. For a string, (which is an object) this ends up not actually modifying the cell contents but at least it doesn't fire an exception. For an integer primitive you would end up with an undefined setter and a noop getter, which would throw exceptions when you try to use undefined as a function to set the cell value.

The fix, as I see it, is to use a regex to strip out the $eval statement from the ng-input in the editable cell template.

I've also cleaned up ng-input.js, added some comments, and added handling
for when the enter key is used to modify a cell's contents