Adjust footer display of total items and filtered items when paging enabled
The footer display of Total Items and Showing Items are correct within the context of all data on the client. However, in the case of server side paging, neither of these counts align with the total data on the server. This commit adjusts the maxRows function in footer.js and the totalFilteredItemsLength function in grid.js to account for server side paging