Bugfix: getInstanceType broken in IE
For some reason, getting an object constructor name as a string in IE adds
a single space at the end.  This was breaking a couple plugin test specs,
and could possibly have been breaking other things as well, in the case
that the plugin needed to be referenced by name.