fix(uiGrid): Fix elm height calc for hidden grids
The header row height could not be calculated when the grid was hidden,
because the absolutely-positioned cells did not count towards the
container height.

To fix this, I made changes to the elementHeight/elementWidth functions
that will, if the element in question is hidden, append the element to the
document body with visibility: hidden.

I also had to switch the cells from being absolutely positioned to
relatively positioned and floated left.  Hopefully this will not have any
further repercussions