WIP: column resizing
Had to alter style computation so they can specify a priority. Also added
a refresh function to the resizer that forces a scroll event so we don't
run into a weird issue where the scrollbar is extending beyond the
viewport and causing offsets between the header canvas scrollLeft and body
canvas scrollLeft.

Right now the offsets don't occur but the horizontal scrollbar DOES extend
3px beyond the right viewport if you scroll all the way to the right and
then resize a column smaller than it was.  Need to fix that.