Added rtl.less
Hacking at RTL support with just CSS so far.

So far the basic layout is fine. Vertical scrolling is fine. And the
layout is fine for horizontal scrolling is fine. Initial clicks on a
horizontal scroll blow away the entire grid. It obviously busts some
calculation out of range. Looking at that part now.