Bug fix for filtered select and de-select
Currently, when selectionBox is enabled with non-empty filter text, the toggleSelectAll is not behaving correctly, which causes inconsistency between selected row number and display of checkboxes of those rows. Because the original code tried to wipe out all other selections whether selectFilter option is true or not, which should not be the case. When selectFilter option is on, invisible rows selection shouldn't be affected.

Easy steps to reproduce:
1. Go to this plucker: http://plnkr.co/edit/gwJMKcNvrnk6FA60T2r7?p=preview
2. Click header checkbox to select All.
3. On the top filter input box, type in "Jacob"
4. Click header checkbox to De-select.
5. Look at footer info stating "Selected Items 0"
6. Clear the filter box.
7. All other rows looked still selected but they are actually not.