bugfix(column resizing): Fixes #1360
Column resizing wasn't accounting for the switch to render containers from a single grid "body" element.
Added a getRenderContainer() method to GridColumn, with test specs
Added GridRenderContainer test spec skeleton, with test to make sure the argument passed for the grid is an actual Grid object
^ That required added a method to gridUtil that returns an object type as a string. Can't use instanceof because it would require a circular dep between GridRenderContainer and Grid.
Had to remove `width: 100%` from cell.less as it was causing column resizing to redraw the cell with the entire width of the window. Didn't appear to break anything else but it may have.