Updates to ngDoc: GridOptions and ColumnDef in each feature
Included an index.ngdoc for the API, providing an overview of how
to interpret the API.  Moved the majority of the GridOptions and
ColumnDef settings back to the individual features, ngDoc doesn't
provide a way to document in both places.  Some wording tidyup and
standardisation of naming - in particular ColumnDef rather than
ColDef everywhere.