Add support for a custom filter.condition function.
Instead of passing a uiGridConstants.filter.CONSTANT or RegExp
as a filter.condition, a user may pass a custom function that
will be run for each row. The function gets the following
arguments passed to it: searchTerm, cellValue, row, column.
It should return true if that row passes the filter, false if
it does not.