Update API docs to reflect filter changes.
Also add a new definition for ui.grid.class:GridOptions.columnDef.filter
to docblocks, however this is not getting picked up by ngDoc. Perhaps
this is something someone else can help me figure out. Or it is not
something that is necessary, in which case we can take it out.