Fix: Sorting with disabled ColMenu throws JS error
If column menu is disabled in options and you click on a column to
sort, this JS erro occurs:

TypeError: Cannot read property 'hideMenu' of undefined
In ui-grid-unstable its in line 904:
uiGridCtrl.columnMenuScope.hideMenu();

'columnMenuScope' is initialised in 'uiGridColumnMenu.link()' which does not
seem to get executed if menu was disabled.  Maybe it's because uiGridColumnMenu
is a directive, and it is only used if menu is enabled.

See #1555, comments