bugfix(Pinning): Pinning styles applied wrong
The custom CSS for pinned containers was not set up right and thus the
canvas and viewports did not have proper sizing. Specifically, the pinned
container canvas size was too small, causing scrollTop to be set to an
incorrect value, and making scrolling look out of sync.

This change fixes the CSS and makes all the viewports and canvas size
properly. I've had to alter the class names a bit. There is no longer a
simple .left or .right. These classes have the property prefix to prevent
any collisions in user apps.