Fix #1595 row_header sometimes not adding correctly
Introduces a new call to `buildColumns`, and adds a name to all
gridColumns (was expected to be present within `buildColumns`)

Also made handleWindowRefresh a publicApi method, as part of this
same issue I was getting issues with column widths that I think
can be resolved by calling this.  Better to have it as a public
method.