Fix #1591 cellNav not scrolling header
This is most of a fix, there is a remaining issue when someone
scrolls rapidly - we're relying on the grid debounce to tell us
if the grid triggered the event or not.  If someone generates two
scroll events within 300ms (say, by holding down the left scroll arrow)
then we'll ignore those events and header still gets out of alignment.