bugfix(GridRenderContainer): bad minimum col calc
In the case that no columns had been rendered yet, there are no drawnWidth
values to use. The minColumnsToRender() method was trying to use them and
ending up with NaN values, causing the minimum col count to be too low.

This caused the grid to only render the absolute minimum number of columns
(i.e. 1 + options.excessColumns). This change makes the method default to
0 for invalid drawnWidths.

Fixes #1497