bugfix(Column Resizing): bad fractional pixels
There were some places where fractional pixel values were showing up and
causing the viewport to be 1 less pixel than the sum of the column widths,
causing the header to wrap.

This change forces the column widths to be integers, with parseInt().
There may be a way to do fractional values across the board but for now
this hopefully fixes the issue and works in all browsers.

Fixes #1619