Fix #1023 (rowSort): allow nulls to propagate to sort functions
Add a new core.api of sortHandleNulls, and use this same method in
each of the sorters.  Add unit tests to sorters.  Minor fixes along with
that to allow false values to be sorted, and to not accidentally fail to
sort 0 and false values.

Extend tutorial to demonstrate sort function.