fix(Column Resizing): Only compile elm if added
The resizeLeft and resizeRight elements were being $compiled whether or
not they were added to the DOM with prepend()/append(). This worked for
some reason with jQuery but without jQuery present angular was failing
with exceptions at calls to insertBefore() because it was looking for
element.parent and an unattached element has no parent.

Fixes #1684, #1655