fix(uiGridUtil): getBorderSize() missing "width"
The correct property name to use is "borderWidth", not just "border".
"border" works in Chrome but was breaking in Firefox.

Also had to change .ui-grid-header's box-sizing to content-box so IE11
would include the border in height calcs. AND finally IE11 was returning
fractional heights so Grid parseInt()s the returned values.