fix(Selection): Remove IE11 ellipsis in header
For some reason IE11 turns the &nbsp; in the selection row header's header
cell into an ellipsis. Adding text-overflow: clip to .ui-grid-contents
fixes this but as a custom headerCellTemplate has to be defined, the
easier thing is just to remove all the unneeded stuff from the header cell
completely.

This change will probably be overwritten if and when we use the header
cell to actually do something.