fix(e2e Tests): Timezone offsets causing failures
The e2e tests for the footer tutorial check for the existence of
date-based text in one of the "max" aggregation columns. This works in
some timezones but fails in others because the tests location can be on
either side of the date line.  I have *commented out* those tests for now.

BREAKING CHANGE: Protractor updated to latest version (1.5.0)

Running `npm install` and `grunt install` should set up the new version.
I had to update the syntax for accessing the protractor instance as well
as it's changed since 1.0.0