fix(uiGridRow): Fix memory leak
Bindings were being attached to the uiGridRow scope but never removed
because the scope was not being destroyed even though child elements were
being added. This was fixed by creating new scopes for uiGridRow when
compiled/re-compiled.

Also, added an element-based `$destroy` handler for ui-grid-cell. It
already had one for the $scope.