fix(Grid): Allow col reordering with column defs
Once initialized, the grid was displaying buggy behavior when swapping
column definitions in or out. This change fixes that behavior, and allows
for Grid.buildColumns() to reorder columns according to the order of
columnDefs by supplying an option parameter with the property
`orderByColumnDefs` set to true.

This also required changing the call to buildColumns() within the grid's
dataWatchFunction so that it uses this option.

Fixes #1948