fix(cellNav): If cellNav api exists, turn on stuff
There was a bug caused by a combination of cellNav and subgrids. If the
parent grid had cellNav turned on but the subgrids didn't, they would
still attach event handlers for cellNav because the directive was looking
for the cellNav controller in its parent chain. It existed, but on the
wrong grid.

This change looks for the presence of the cellNav api in the Grid to turn
on the event handlers.

Fixes #2294