Fix plunker: border-box style on headercell
For me (but not for everyone) in plunker I'm getting the headers
wrapping.  The cause seems to be that the rightmost header cell
border doesn't have any space - so the whole header row
is 1px too wide.  No idea why that happens to me and nobody
else (well, not @swalters or @c0bra).

The header apparently needs content-box for height, aiming to
put border-box on the header-cell, which might sort things out.

This needs verifying on IE11, because the content-box change
was to fix IE11.