fix(uiGridHeader): Fix header height growth bug
Header cells were growing in height due to a calculation bug where the
grid was reading the height of the entire header element rather than just
the repeater where the header cells were contained.  That means that a
custom header with extra stuff in it would cause headers to grow in height
on each canvas refresh.

This change splits header height calculations in two: one for the headers
themselves, which keeps headers between render containers consistent in
height, and another for header cells which uses the header canvas height
to keep header cells the same height.

Fixes #2781