fix(uiGridRenderContainer): Use header min-height
Setting a fixed height on the header cells prevented re-calculating
heights later on, if the column defs were provided async. So if you
enabled filters they would not appear because the header heights would be
set too low.

Changing from height to min-height lets the headers grow as needed.

Fixes #2768