fix(Pinning): Move rowStyle() to uiGridViewport
rowStyle() in GridRenderContainer was setting both row AND column offsets.
The problem with that is pinned containers reference a different render
container for their rows than their columns. Using the column
margin-offset from the body container resulted in render issues.

This change moves the rowStyle() function to uiGridViewport's controller,
and it uses the rowContainer for the row offset and the colContainer for the
column offset.

Fixes #2821