Fix (columnMenu): Fix #2852 positioning of column menu once scrolled
The column menus for columns to the right were in the wrong place. This
was due to the change to the header to use a table format, which also
appeared to include using offsetLeft on the parent to reflect virtualised
columns to the left.  Take this offsetLeft into account as well in the
column menu positioning calculations.