Fix (exporter): Fix #2313 export for IE9
Information and code provided by @alexandery.  I have no ability to test
it, but it should only impact IE9, which apparently doesn't work otherwise
anyway.

The main concern here is that this creates and destroys an iFrame in IE9,
and that iFrame apparently messes up formatting, so requires application of
width: auto !important; on the ui-grid-viewport.  That appears to break formatting
in chrome on all features tutorial, so I haven't applied it.  Presumably that means
formatting will be ugly in IE9.