Feat(grouping): alpha release of grouping feature
This is a basic grouping function, providing a basis on which we
can extend and debug.  It provides:
- multilevel grouping
- aggregations on any non-grouped columns
- menu items to add/remove grouping
- menu items to add/remove aggregations
- API for manipulating grouping and rows
- groupingRowHeader to expand and collapse grouping.

The code builds on the sort logic, and heavily on the rowProcessor logic.
It adds internal rows to the rendered rows (but not grid.rows), these
are regenerated each time the rowProcessor runs.

It maintains expanded/collapsed state in grid.grouping.rowExpandedStates,
so that this persists across row builds.

It should interact cleanly with filtering and sorting.  It doesn't interact
well with edit and probably some of the other features.  It doesn't work well
with column bindings that are complex, and that needs some more thought - behind
the scenes it's creating a fake entity, and writing into that entity our aggregation
values.  If the fieldName isn't one we can easily deal with then writing to the entity
is hard.

This PR also includes a change to fontello config to restore fonts, and some changes
in core:

- changes to the way we handle invisible rows - include a new invisibleReason to replace
  the old forceInvisible, and changed logic to use it
- changes to suppressRemoveSort, which stops sorts from being messed up - it always used
  colDef, changed to copy down onto col.suppressRemoveSort and use from there, using this
  logic to stop the sort being messed up for grouped columns
- added a new queueGridRefresh() method to debounce grid refreshes.  Was using a lot of them.

More details are included in the tutorial, and some design information in the top of the code
file and therefore also in the root of the API:grouping documentation.