deep watch on filter.term, enable data structures as column terms
This is part of #2918 to make it possible to use arbitrary complex
custom column filters.

A use case is a date range picker with a term object like
{from: Date, to: Date} or a multiselect.

There should be no measurable performance impact for normal text
filters. Instead of comparing term directly with === angular now
calls angular.equals which starts with

  function equals(o1, o2) {
    if (o1 === o2) return true;