Enh(core): add doco to refresh methods, move to queueGridRefresh
Make queueRefresh (which calls refreshCanvas) and queueGridRefresh
(which calls grid.refresh) available on the api.  Document when
to use which, and what they do.

Change the core application to use queueGridRefresh in preference
to refresh, which means we get a debounce if this was being called
multiple times in a digest - for example once for each column.

Didn't apply to areas that appeared to be waiting on the promise
that it returns - e.g. didn't apply to pinning.