Fix(touch events): touch events on resize, move, header click/sort/menu, selection
Fix #3084, #2838, #2616, #2550, #2303.  All relate to handling of events
on touch devices.  General philosophy is to register down events for both
touch and mouse, but once we get a down event, we only listen for move and
up events of the type we got a down on.  This is a similar approach to
hammer.js.

Also includes a big refactor of how move columns listens for events, makes
it more like other handlers.

This all works for me on iPad, and on OSX/chrome (non-touch).  I don't have
the ability to test a machine that has both mouse and touch, so waiting on
someone else to look at that.