Fix(resize): Fix #2550 - first sample code
Try to address this issue by setting events for both touch and mouse, but
once a touchdown (or mousedown) event starts, setting only the associated
touch (or mouse) move and up handlers, and disabling the opposite down
handler until this event completes.  Similar approach to hammer.js.

Tests suppressed whilst we see whether this works on a touch device,
if it does then we can work out why this breaks the unit tests.