Fix(infiniteScroll): fix #2730, fix #2827, BREAKING changes
Substantial refactor of infinite scroll.  Provides new functionality, a much more
detailed tutorial, and removes the code that crept into core in #2730 (which I merged
when I shouldn't have).  Clarifies the handling, removes a viewPort directive that was
unnecessary, makes the handling for scroll up and down consistent.

Strictly speaking the API hasn't changed, but there are some new settings and the
default behaviour has subtly changed in that scrollUp isn't enabled by default.  This
should be more logical to users, but probably does require a review of the tutorial and API.