Enh(infiniteScroll): BREAKING: change percentage to rowsToEnd
Infinite scroll was getting more data at a percentage from the end.  This meant that as you
got more pages you would trigger the request for more data further and further from the end,
and eventually a request for new data would immediately trigger a request for still more data, as
your scroll kept moving too close to the end.

Change instead to infiniteScrollRowsFromEnd, which doesn't have this problem.