fix(uiGridHeader): Allow header to shrink in size
The combo of ui-grid-selection and filtering was causes the headers to be
unable to shrink in size if filtering were toggled off, because the
selection grid header had an explicit height and when we saw it we would
use that as the height to match up with.

This change removes all explicit heights from the "max height" check, so
we only rely on rendered dynamic heights to calculate the max height.

A unit test has also been added to the selection specs to cover this case.

Fixes #3138