feat(saveState): add pinning to save state
The idea was to update saveState to include pinning. This turned
out to be more involved than initial anticipated. I ended up
having to update the pinning feature to include a public event
and method so saveState could include pinning as an option.

I updated documentation as I worked through it and even updated the
tutorial to include pinning. This is in reference to an feature
request I made recently #3123.