fix(uiGrid): Wait for grid to get dimensions
If the grid has no width initially, wait for up to 2 seconds for the grid
to be shown and have dimensions. Once that happens it will re-initialize
its height and width and redraw the canvas.

This should fix the problems with the grid taking up too much space in a
modal.

BREAKING CHANGE: gridUtil will no longer calculate dimensions of hidden
elements