Change expandable row height according to its expandedRowHeight property
Added "expandedRowHeight" property to the expanded row.
If it is not defined, the grid's default is used.
Changed the template files to use the "expandedRowHeight" instead of the
grid's default.

Added "rowExpandedBeforeStateChanged" event to "expandable" feature

Can now hook to the event before the change. One scenario this can be
helpful is to dynamically change the expandedRowHeight