Fix(scrolling): dreadful iPad performance, badly formed
A handful of issues, mostly comes down to poor use of $interval in the throttle method (my
change), and also fixing a defect where the aggregations weren't calculated by turning the
throttle into a trailing throttle.  This resulted in the grid waiting 0.5s before running the
aggregation, but the aggregation triggered a $digest since a bound value had changed, and in turn
that called aggregation again.

Moved the aggregation call to be triggered by the rowsRendered event.  This event is emitted
whenever visible rows change, which should suit aggregation.  This meant I had to move the
call to aggregation into the gridFooterCell, as I needed a destroy event in order to destroy the
registered api listener.