Altered how SELECT filters compare values
Since searchOption now uses "track by" instead of "selectAs", the
entire searchOption object will be set as the term.  This would require
a custom condition function for every SELECT filter since you would
need to extract the term.value before comparing to the cell value.
This code extracts the .value before it does the comparison.  It also
uses angular.equals() to compare the SELECT filters by default instead
of trying to use a starts with regex.  This solves the problem of using
objects instead of strings for searchOption values.