feat(grouping): Add support for multiple column instances
This changes the way aggregated values are stored on the row, and introduces
an expression function to retreive the values. This should have perfomance
advantages over a custom a template, but does have the consequence of adding
 a property to every row entity, which holds a reference to that function.
 While we don't change any of the of the existing properties, it could be
 undesireable to some users that we "alter" their object.

Change the logic in the average aggregation function, to allow for more
complete cleanup

Throw error on incompatible option. The grouping feature does not work
with flatEntityAccess, so it will now throw an error if the user passes
that option.

 Closes #3435