Enh(treeView): BREAKING treeView uses treeBase, will eventually share with grouping
Write a new library for treeBase, which will offer features to both treeView and grouping.
Implement into that library recursive sorting, filtering where the parent nodes are shown
if any children are filtered in, aggregations, and all the expand/collapse logic. Implement
better checking of expand all so it knows if you've manually expanded all the rows.

Provide a new tutorial 319 complex tree, demonstrating new features.

Upgrade treeView to use this library.  BREAKING changes are that many of the service
calls and constants are now on treeBase instead of treeView, otherwise behaviour is
similar.  Thought about providing a deprecation, but given this is an alpha feature
in a release candidate grid, figured better not to create the cruft.

Made some ancillary changes in core:
- getCellValue has the ability to return entity data based on $$col.uid, this masks
  the entity values if present (one extra if in getCellValue)
- rowSearcher returns all rows and sets visible flag only (it was only returning visible rows).
  It used to do this, I changed it for performance reasons (no need to sort invisible rows), however
  treeView needs to get the invisible rows as well, as it turns some of them
  back into visible.  Compromise I guess
- rowSorter has a new flag - ignoreSort.  Does as you might expect.