enh(grouping): BREAKING share code from treeBase, changes some interfaces
Migrate grouping to using treeBase for the underlying tree.  Gives more
consistency of behaviour, and enables more features.  Fixes the problem
where grouping didn't work on columns with complex bindings.  Allows
fine tuning of aggregations through customAggregationFinalizerFn and
customAggregationFn.  Allows group column to have aggregations other
than count.  Allows cellTemplates and code to get at the individual
parts of the aggregation and groupHeader in a cell - can access the
label part and the value part separately to allow formatting and filtering
as necessary.

Breaking changes are:
- methods such as expandAll are now on treeBase, not on grouping
- aggregations move from being col.grouping.aggregation to col.treeAggregation
- saveState format changes, if you persisted that across instantiations your
  saved states may not reload cleanly
- groupLevel is now treeLevel
- groupingSuppressAggregationText is now handled via customTreeAggregationFinalizerText
- groupingRowHeaderWidth is now treeRowHeaderWidth
- groupingIndent is now treeIndent
- groupingRowHeaderAlwaysVisible is now treeRowHeaderAlwaysVisible, and default has changed to true

Includes minor fixes on treeBase and treeView.